Move to AnnotationContexts to mod.rs, nuke utils.rs
