+++ compiler/annotation/function.rs
@@ -36,12 +36,11 @@ use typeql::{
 
 use crate::{
     annotation::{
-        FunctionAnnotationError, TypeInferenceError,
+        AnnotationContext, FunctionAnnotationError, PipelineAnnotationContext, TypeInferenceError,
         pipeline::{
             AnnotatedStage, RunningVariableAnnotations, annotate_pipeline_stages, resolve_reducer_by_value_type,
         },
         type_seeder,
-        utils::{AnnotationContext, PipelineAnnotationContext},
     },
     executable::reduce::ReduceInstruction,
 };
