+++ executor/tests/execute_isa.rs
@@ -13,8 +13,8 @@ use answer::variable::Variable;
 use compiler::{
     self, ExecutorVariable, VariablePosition,
     annotation::{
-        function::EmptyAnnotatedFunctionSignatures, match_inference::infer_types_for_test_only,
-        utils::PipelineAnnotationContext,
+        PipelineAnnotationContext, function::EmptyAnnotatedFunctionSignatures,
+        match_inference::infer_types_for_test_only,
     },
     executable::{
         function::ExecutableFunctionRegistry,
